# packages/api-client/openapi/workforce-api.yaml
openapi: 3.1.0
info:
  title: WORKFORCE API
  version: 1.1.0
  description: "Canonical Laravel API contract for Workforce ERP.


    User-facing business endpoints authenticate with Laravel Sanctum bearer

    tokens. `X-API-TOKEN` is reserved for explicitly scoped trusted

    server-to-server endpoints and is not a replacement for user authorization.

    "
servers:
  - url: http://localhost:8000
    description: Local Laravel API
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      operationId: login
      summary: Sign in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/password/forgot:
    post:
      tags:
        - Authentication
      operationId: forgotPassword
      summary: Request a password reset link
      description:
        Returns the same accepted response for unknown or ineligible accounts. Delivery uses the configured Laravel
        mail transport.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        type: string
                        example: If the account is eligible and email delivery is available, a password reset link will arrive shortly.
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/password/reset:
    post:
      tags:
        - Authentication
      operationId: resetPassword
      summary: Reset a password with the Laravel password-broker token
      description: On success, changes the password and revokes all existing Sanctum tokens for the account.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/otp/request:
    post:
      tags:
        - Authentication
      operationId: requestOtp
      summary: Request a one-time sign-in code
      description: Returns the same accepted response for unknown or blocked accounts to avoid disclosing account existence/state.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        type: string
                        example: If the account is eligible and email delivery is available, a one-time code will arrive shortly.
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/otp/verify:
    post:
      tags:
        - Authentication
      operationId: verifyOtp
      summary: Verify a one-time code and issue a Sanctum token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - email
                - code
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  pattern: ^\\d{6}$
                  minLength: 6
                  maxLength: 6
                client:
                  type: string
                  enum:
                    - portal
                    - admin
                  default: portal
      responses:
        "200":
          description: One-time code verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/sso/redirect/{provider}:
    get:
      tags:
        - Authentication
      operationId: getSsoRedirect
      summary: Create an SSO authorization redirect URL, state, and server-bound PKCE transaction
      security: []
      parameters:
        - $ref: "#/components/parameters/SsoProvider"
      responses:
        "200":
          description: Redirect information generated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - success
                  - redirect_url
                  - state
                properties:
                  success:
                    type: boolean
                    const: true
                  redirect_url:
                    type: string
                    format: uri
                  state:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          description: SSO provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/sso/callback/{provider}:
    post:
      tags:
        - Authentication
      operationId: completeSso
      summary: Complete an SSO authorization-code + PKCE flow
      description: The API consumes the one-time state and sends its cached PKCE verifier to the selected provider token endpoint.
      security: []
      parameters:
        - $ref: "#/components/parameters/SsoProvider"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - code
                - state
              properties:
                code:
                  type: string
                  maxLength: 4096
                state:
                  type: string
                  minLength: 48
                  maxLength: 48
      responses:
        "200":
          description: SSO authentication completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "419":
          description: Invalid or expired SSO state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          description: SSO provider is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: SSO provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/me:
    get:
      tags:
        - Authentication
      operationId: getCurrentUser
      summary: Return the authenticated user context
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                  - success
                  - user
                properties:
                  success:
                    type: boolean
                    const: true
                  user:
                    $ref: "#/components/schemas/AuthUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/sessions:
    get:
      tags:
        - Authentication
      operationId: listAuthSessions
      summary: List active browser sessions or API-token sessions for the authenticated client
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Active token sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/sessions/{tokenId}:
    delete:
      tags:
        - Authentication
      operationId: revokeAuthSession
      summary: Revoke one of the authenticated user’s sessions
      security:
        - BearerAuth: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/auth/password/change:
    post:
      tags:
        - Authentication
      operationId: changePassword
      summary: Change the authenticated user’s password
      description: Requires the current password and revokes all Sanctum tokens after a successful change.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      operationId: logout
      summary: Log out the current browser session or revoke the current API token
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current token revoked
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Logged out successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout-all:
    post:
      tags:
        - Authentication
      operationId: logoutAll
      summary: Revoke every Sanctum API token for the authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: All sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users:
    get:
      tags:
        - Users
      operationId: listUsers
      summary: List users visible to the current owner/admin
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: organization_id
          in: query
          schema:
            type: integer
            minimum: 1
        - name: role
          in: query
          schema:
            oneOf:
              - $ref: "#/components/schemas/UserRole"
              - type: string
                const: all
        - name: status
          in: query
          schema:
            oneOf:
              - $ref: "#/components/schemas/UserStatus"
              - type: string
                const: all
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags:
        - Users
      operationId: inviteUser
      summary: Invite a user into an organization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserRequest"
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/users/options/organizations:
    get:
      tags:
        - Users
      operationId: listUserOrganizationOptions
      summary: List organizations the current user can administer
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Organization options
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrganizationOption"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/options/roles:
    get:
      tags:
        - Users
      operationId: listUserRoleOptions
      summary: List assignable coarse-grained user roles
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Role options
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RoleOption"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/options/employees:
    get:
      tags:
        - Users
      operationId: listUserEmployeeOptions
      summary: List employees available for user linking
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Employee options
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/EmployeeOption"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/{user}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags:
        - Users
      operationId: getUser
      summary: Get a user in a manageable shared organization
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Users
      operationId: updateUser
      summary: Update a user membership/profile link
      description: Partial-update semantics are supported even though the route uses PUT for frontend compatibility.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/users/{user}/activate:
    patch:
      tags:
        - Users
      operationId: activateUser
      summary: Activate a user membership
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user}/deactivate:
    patch:
      tags:
        - Users
      operationId: deactivateUser
      summary: Deactivate a user membership and revoke the user's tokens
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user}/suspend:
    patch:
      tags:
        - Users
      operationId: suspendUser
      summary: Suspend a user membership and revoke the user's tokens
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{user}/resend-invitation:
    post:
      tags:
        - Users
      operationId: resendUserInvitation
      summary: Re-send a user's invitation email
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Invitation delivery attempted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: object
                        required:
                          - delivered
                          - sent_at
                        properties:
                          delivered:
                            type: boolean
                          sent_at:
                            anyOf:
                              - type: string
                                format: date-time
                              - type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/employees:
    get:
      tags:
        - Employees
      operationId: listEmployees
      summary: List employees in the authenticated user's active organizations
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: department
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated employees
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Employee"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/employees/options:
    get:
      tags:
        - Employees
      operationId: getEmployeeDirectoryOptions
      summary: Get employee-directory filter options
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Department and location values
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: "#/components/schemas/EmployeeDirectoryOptions"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/employees/summary:
    get:
      tags:
        - Employees
      operationId: getEmployeeDirectorySummary
      summary: Get employee-directory summary counts
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Employee summary counts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: "#/components/schemas/EmployeeDirectorySummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Laravel Sanctum personal access token
      description: Personal access token for explicitly non-browser API clients. Browser sign-in uses Sanctum session cookies and does not return a bearer token.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-TOKEN
      description: Environment-configured shared token used only by explicitly scoped trusted service endpoints.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not authorized for the operation/resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    TooManyRequests:
      description: Rate limit or workflow attempt limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
        message:
          type: string
...
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        client:
          type: string
          enum:
            - portal
            - admin
          default: portal
    AuthSuccessResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - user
      properties:
        success:
          type: boolean
          const: true
        user:
          $ref: "#/components/schemas/AuthUser"
...
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 255
...